update endoint voucher vehicle status change and handle from upload file

+++ templates/resources/midleware/MD_SGI.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+from templates.controllers.vouchers.vouchers_controller import (
+    update_voucher_vehicle_status,
+)
 from static.constants import log_file_sgi_chv
 from templates.misc.Functions_Files import write_log_file
 from templates.Functions_Utils import create_notification_permission_notGUI
@@ -740,6 +743,7 @@ def update_voucher_vehicle_api(data, data_token):
         data["type"],
         data["received_by"],
         data.get("observations"),
+        data.get("status"),
     )
     if not flag:
         return {
@@ -799,20 +803,49 @@ def update_voucher_vehicle_api(data, data_token):
 
 
 def delete_voucher_vehicle_api(data, data_token):
-    flag, error, result = delete_voucher_item(data["id"])
+    time_zone = pytz.timezone(timezone_software)
+    time_older = datetime.now(pytz.utc).astimezone(time_zone) - timedelta(days=365)
+    timestamp = datetime.now(pytz.utc).astimezone(time_zone).strftime(format_timestamps)
+    flag, error, voucher_data = get_vouchers_vehicle_with_items(
+        time_older, id_voucher=data["id"]
+    )
     if not flag:
         return {
             "data": None,
-            "msg": "Error at deleting vehicle voucher",
+            "msg": "Error at getting vehicle voucher by id",
             "error": str(error),
         }, 400
-    flag, error, result = delete_voucher_vehicle(data["id"])
-    if not flag:
+    if not (isinstance(voucher_data, list) or isinstance(voucher_data, tuple)):
         return {
             "data": None,
-            "msg": "Error at deleting vehicle voucher",
-            "error": str(error),
+            "msg": "Error at getting vehicle voucher by id: result is not a list or tuple",
+            "error": str(voucher_data),
         }, 400
+    status = voucher_data[20]
+    if status == 0:  # delete if was not signed
+        flag, error, result = delete_voucher_item(data["id"])
+        if not flag:
+            return {
+                "data": None,
+                "msg": "Error at deleting vehicle voucher",
+                "error": str(error),
+            }, 400
+        flag, error, result = delete_voucher_vehicle(data["id"])
+        if not flag:
+            return {
+                "data": None,
+                "msg": "Error at deleting vehicle voucher",
+                "error": str(error),
+            }, 400
+    else:  # cancel if was signed
+        flag, error, result = update_voucher_vehicle_status(3, data["id"])
+        if not flag:
+            return {
+                "data": None,
+                "msg": "Error at canceling vehicle voucher",
+                "error": str(error),
+            }, 400
+
     time_zone = pytz.timezone(timezone_software)
     timestamp = datetime.now(pytz.utc).astimezone(time_zone).strftime(format_timestamps)
 
@@ -823,7 +856,9 @@ def delete_voucher_vehicle_api(data, data_token):
             "type": 2,
             "timestamp": timestamp,
             "user": data_token.get("emp_id"),
-            "comment": "Voucher vehicular eliminado",
+            "comment": "Voucher vehicular eliminado"
+            if status == 0
+            else "Voucher vehicular cancelado",
         }
     )
 
@@ -836,13 +871,22 @@ def delete_voucher_vehicle_api(data, data_token):
             "msg": "Error at deleting vehicle voucher",
             "error": str(error),
         }, 400
-
+    msg = (
+        f"Voucher vehicular eliminado correctamente por el empleado {data_token.get('emp_id')}"
+        if status == 0
+        else f"Voucher vehicular cancelado correctamente por el empleado {data_token.get('emp_id')}"
+    )
+    create_notification_permission_notGUI(
+        msg, ["administracion", "operaciones", "sgi"], data_token.get("emp_id"), 0
+    )
+    write_log_file(log_file_sgi_chv, msg)
     return {"data": [rows_changed], "msg": "Vehicle voucher deleted successfully"}, 200
 
 
 def create_voucher_vehicle_attachment_api(data, data_token):
     """{"filepath": filepath_download, "filename": filename}, data_token"""
     filename = data["filename"]
+
     id_voucher_name = filename.split("-")[0]
     try:
         if (
@@ -925,6 +969,14 @@ def create_voucher_vehicle_attachment_api(data, data_token):
             return {"data": None, "msg": f"Access denied to bucket: {bucket_name}"}, 400
         else:
             return {"data": None, "msg": f"AWS error: {str(e)}"}, 400
+    msg = f"Archivo adjunto agregado: {filename} al voucher {data['id_voucher']} por el empleado {data_token.get('emp_id')}"
+    status = voucher_data[21]
+    if "firma-realizado" in filename.lower():  # if is sign file change status to 1
+        status = 1
+        msg += " y estado actualizado a (firmado)"
+    if "firma-recibido" in filename.lower():  # if is sign file change status to 1
+        status = 2
+        msg += " y estado actualizado a (aprobado)"
     history.append(
         {
             "id_voucher": data["id_voucher"],
@@ -944,7 +996,7 @@ def create_voucher_vehicle_attachment_api(data, data_token):
     )
     extra_info["files"] = files
     flag, error, rows_updated = update_voucher_vehicle_files(
-        data["id_voucher"], history, extra_info
+        data["id_voucher"], history, status, extra_info
     )
     if not flag:
         return {
@@ -952,7 +1004,6 @@ def create_voucher_vehicle_attachment_api(data, data_token):
             "msg": "Error at updating history voucher but file uploaded",
             "error": str(error),
         }, 400
-    msg = "File uploaded successfully but error at updating history voucher"
     create_notification_permission_notGUI(
         msg, ["administracion", "operaciones", "sgi"], data_token.get("emp_id"), 0
     )
